Fixed some bugs in IF, but still not working.
Oddity: the Covariance matrix does not change when running jerk car.
Nor does the velocity or position, only the acceleration changes and the
results are completely incoherent with the vanilla KF. In addition, this
only seems to work with the x0 and P0 of the example (using nil vector
and matrix does not lead to that at all).