Lots of debugging messages: I now now that the issue is not a shift but that PropUntil stalls.
This is a strange issue: if measurements are separated by exactly the time step, then it will stall and not propagate at all. If add a full time step, it'll propagate *a bit* and then stall until there is a large delta between the previous and the following step. Weird af.